pyproject.toml
[project]
name = "my-project"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
	"google-generativeai>=0.8.4",
	"mss>=9.0.1",
	"pillow>=10.4.0",
	"python-dotenv>=1.0.1",
	"pystray>=0.19.5",
]
